reorganizes converter folder structure and names
