+++ DUI3-DX/Converters/Autocad/Speckle.Converters.AutocadShared/ToSpeckle/Geometry/PointToSpeckleConverter.cs
@@ -3,15 +3,15 @@ using Speckle.Converters.Common;
 using Speckle.Core.Models;
 using Autodesk.AutoCAD.DatabaseServices;
 
-namespace Speckle.Converters.Autocad.Geometry;
+namespace Speckle.Converters.Autocad.ToSpeckle.Geometry;
 
 [NameAndRankValue(nameof(DBPoint), NameAndRankValueAttribute.SPECKLE_DEFAULT_RANK)]
-public class DBPointToSpeckleConverter : IHostObjectToSpeckleConversion, IRawConversion<DBPoint, SOG.Point>
+public class PointToSpeckleConverter : IHostObjectToSpeckleConversion
 {
   private readonly IRawConversion<AG.Point3d, SOG.Point> _pointConverter;
   private readonly IConversionContextStack<Document, UnitsValue> _contextStack;
 
-  public DBPointToSpeckleConverter(
+  public PointToSpeckleConverter(
     IRawConversion<AG.Point3d, SOG.Point> pointConverter,
     IConversionContextStack<Document, UnitsValue> contextStack
   )
@@ -23,19 +23,4 @@ public class DBPointToSpeckleConverter : IHostObjectToSpeckleConversion, IRawCon
   public Base Convert(object target) => RawConvert((DBPoint)target);
 
   public SOG.Point RawConvert(DBPoint target) => _pointConverter.RawConvert(target.Position);
-}
-
-public class PointToSpeckleConverter : IRawConversion<AG.Point3d, SOG.Point>
-{
-  private readonly IConversionContextStack<Document, UnitsValue> _contextStack;
-
-  public PointToSpeckleConverter(IConversionContextStack<Document, UnitsValue> contextStack)
-  {
-    _contextStack = contextStack;
-  }
-
-  public Base Convert(object target) => RawConvert((AG.Point3d)target);
-
-  public SOG.Point RawConvert(AG.Point3d target) =>
-    new(target.X, target.Y, target.Z, _contextStack.Current.SpeckleUnits);
 }
