+++ DUI3-DX/Converters/Autocad/Speckle.Converters.AutocadShared/ToSpeckle/Raw/IntervalToSpeckleRawConverter.cs
@@ -1,0 +1,11 @@
+using Speckle.Converters.Common.Objects;
+using Speckle.Core.Models;
+
+namespace Speckle.Converters.Autocad.ToSpeckle.Raw;
+
+public class IntervalToSpeckleConverter : IRawConversion<AG.Interval, SOP.Interval>
+{
+  public Base Convert(object target) => RawConvert((AG.Interval)target);
+
+  public SOP.Interval RawConvert(AG.Interval target) => new(target.LowerBound, target.UpperBound);
+}
