+++ DUI3-DX/Converters/Autocad/Speckle.Converters.AutocadShared/ToSpeckle/Raw/PlaneToSpeckleRawConverter.cs
@@ -3,16 +3,15 @@ using Speckle.Converters.Common;
 using Speckle.Converters.Common.Objects;
 using Speckle.Core.Models;
 
-namespace Speckle.Converters.Autocad.Geometry;
+namespace Speckle.Converters.Autocad.ToSpeckle.Raw;
 
-[NameAndRankValue(nameof(AG.Plane), NameAndRankValueAttribute.SPECKLE_DEFAULT_RANK)]
-public class PlaneToSpeckleConverter : IHostObjectToSpeckleConversion, IRawConversion<AG.Plane, SOG.Plane>
+public class PlaneToSpeckleRawConverter : IHostObjectToSpeckleConversion
 {
   private readonly IRawConversion<AG.Vector3d, SOG.Vector> _vectorConverter;
   private readonly IRawConversion<AG.Point3d, SOG.Point> _pointConverter;
   private readonly IConversionContextStack<Document, UnitsValue> _contextStack;
 
-  public PlaneToSpeckleConverter(
+  public PlaneToSpeckleRawConverter(
     IRawConversion<AG.Vector3d, SOG.Vector> vectorConverter,
     IRawConversion<AG.Point3d, SOG.Point> pointConverter,
     IConversionContextStack<Document, UnitsValue> contextStack
